feat: add eval harness for 500-question MCP coverage scoring (#37)
Refactors monolithic eval.py into modular eval/ package with separate
modules for parsing, running, judging, reporting, and updating coverage.
Run via `python -m eval` with --quick, --category, --limit, --threshold,
--concurrency, and --no-update flags.
