+++ eval/__init__.py
@@ -1,0 +1,2 @@
+"""Eval harness for Qualys MCP Server — sends questions from docs/questions.md
+against the live MCP server via Claude API and scores response quality."""
